+++ protocols/shuffling_protocol.go
@@ -77,7 +77,7 @@ type ShufflingProtocol struct {
 	TargetOfShuffle   *[]libunlynx.ProcessResponse
 
 	CollectiveKey kyber.Point //only use in order to test the protocol
-	Proofs        int
+	Proofs        int64
 	Precomputed   []libunlynx.CipherVectorScalar
 
 	Query *libdrynx.SurveyQuery
